requirements.txt
aiogram
loguru
python-dotenv
tortoise-orm
aiosqlite
aerich
tomlkit
